Remove P2PNodesManager and update test to test NodeConnectionManager fully
