+++ WalletWasabi/Services/NodesManagement/NodeConnectionManager.cs
@@ -1,3 +1,7 @@
+using NBitcoin;
+using NBitcoin.Protocol;
+using WabiSabi.Crypto.Randomness;
+using NBitcoin.Protocol.Behaviors;
 using System.Collections.Concurrent;
 using System.Collections.Generic;
 using System.Collections.Immutable;
@@ -6,9 +10,7 @@ using System.Linq;
 using System.Net;
 using System.Threading;
 using System.Threading.Tasks;
-using NBitcoin;
-using NBitcoin.Protocol;
-using NBitcoin.Protocol.Behaviors;
+using WalletWasabi.Extensions;
 using WalletWasabi.Helpers;
 using WalletWasabi.Logging;
 using static WalletWasabi.Services.Workers;
@@ -18,6 +20,16 @@ namespace WalletWasabi.Services.NodesManagement;
 public interface INodesRegistry
 {
 	ImmutableArray<Node> Nodes { get; }
+
+	Task<Node> GetNodeForSingleUseAsync(CancellationToken cancellationToken);
+
+	void DisconnectNode(Node node, string reason);
+
+	void DisconnectNodeIfEnoughPeers(Node node, string reason);
+
+	double GetCurrentTimeout();
+
+	void UpdateTimeout(bool increaseDecrease);
 }
 
 public class NodeConnectionManager : INodesRegistry, IDisposable
@@ -31,8 +43,8 @@ public class NodeConnectionManager : INodesRegistry, IDisposable
 	private static readonly TimeSpan QuickDisconnectThreshold = TimeSpan.FromSeconds(30);
 	private static readonly TimeSpan MaintainInterval = TimeSpan.FromSeconds(6);
 	private static readonly TimeSpan RotateInterval = TimeSpan.FromMinutes(3);
-	private static readonly TimeSpan CrawlerConnectionTimeout = TimeSpan.FromSeconds(10);
-	private static readonly TimeSpan CrawlerHarvestTimeout = TimeSpan.FromSeconds(3);
+	private static readonly TimeSpan CrawlerConnectionTimeout = TimeSpan.FromSeconds(60);
+	private static readonly TimeSpan CrawlerHarvestTimeout = TimeSpan.FromSeconds(15);
 
 	private readonly Network _network;
 	private readonly List<NodeBehavior> _templateBehaviors = [];
@@ -52,6 +64,10 @@ public class NodeConnectionManager : INodesRegistry, IDisposable
 	private int _isReevaluating;
 	private DateTimeOffset _lastMaintainTime;
 	private DateTimeOffset _lastRotateTime;
+
+	private int _timeoutsCounter;
+	private int _currentTimeoutSeconds = 16;
+
 	private bool _isDisposed;
 
 	public NodeConnectionManager(
@@ -155,6 +171,106 @@ public class NodeConnectionManager : INodesRegistry, IDisposable
 		{
 			Interlocked.Exchange(ref _isReevaluating, 0);
 		}
+	}
+
+	public async Task<Node> GetNodeForSingleUseAsync(CancellationToken cancellationToken)
+	{
+		while (!cancellationToken.IsCancellationRequested)
+		{
+			if (Nodes.Length == 0)
+			{
+				await Task.Delay(TimeSpan.FromMilliseconds(100), cancellationToken).ConfigureAwait(false);
+				continue;
+			}
+
+			var nodes = Nodes;
+			var node = nodes.RandomElement(SecureRandom.Instance);
+
+			if (node is not null && node.IsConnected)
+			{
+				return node;
+			}
+			Logger.LogTrace($"Selected node is null or disconnected.");
+
+			await Task.Delay(10, cancellationToken).ConfigureAwait(false);
+		}
+
+		cancellationToken.ThrowIfCancellationRequested();
+		throw new InvalidOperationException("Failed to retrieve a connected node.");
+	}
+
+	public void DisconnectNodeIfEnoughPeers(Node node, string reason)
+	{
+		// Always keep at least 5 nodes connected
+		if (Nodes.Length <= 5)
+		{
+			return;
+		}
+
+		if (node.SupportsCompactFilters)
+		{
+			return;
+		}
+
+		DisconnectNode(node, reason);
+	}
+
+	public void DisconnectNode(Node node, string reason)
+	{
+		Logger.LogInfo(reason);
+		node.DisconnectAsync(reason);
+	}
+
+	public double GetCurrentTimeout()
+	{
+		// More permissive timeout if few nodes are connected to avoid exhaustion.
+		return Nodes.Length < 3
+			? Math.Min(_currentTimeoutSeconds * 1.5, 600)
+			: _currentTimeoutSeconds;
+	}
+
+	/// <summary>
+	/// Current timeout used when downloading a block from the remote node. It is defined in seconds.
+	/// </summary>
+	public void UpdateTimeout(bool increaseDecrease)
+	{
+		if (increaseDecrease)
+		{
+			_timeoutsCounter++;
+		}
+		else
+		{
+			_timeoutsCounter--;
+		}
+
+		var timeout = _currentTimeoutSeconds;
+
+		// If it times out 2 times in a row then increase the timeout.
+		if (_timeoutsCounter >= 2)
+		{
+			_timeoutsCounter = 0;
+			timeout = (int)Math.Round(timeout * 1.5);
+		}
+		else if (_timeoutsCounter <= -3) // If it does not time out 3 times in a row, lower the timeout.
+		{
+			_timeoutsCounter = 0;
+			timeout = (int)Math.Round(timeout * 0.7);
+		}
+
+		// Sanity check
+		var minTimeout = _network == Network.Main ? 3 : 2;
+
+		if (timeout < minTimeout)
+		{
+			timeout = minTimeout;
+		}
+		else if (timeout > 600)
+		{
+			timeout = 600;
+		}
+
+		_currentTimeoutSeconds = timeout;
+		Logger.LogInfo($"Current timeout value used on block download is: {timeout} seconds.");
 	}
 
 	private async Task<PeerInfo[]> GetPeersAsync(CancellationToken cancellationToken)
@@ -593,10 +709,13 @@ public class NodeConnectionManager : INodesRegistry, IDisposable
 				discoveredAt: now,
 				lastSeen: now);
 
+			Logger.LogDebug($"Connected to endpoint '{endpoint}'");
 			_discoveryCoordinator?.Post(new PeerDiscoveredMessage(peer));
 		}
-		catch
+		catch (Exception e)
 		{
+			Logger.LogTrace($"Failed to connect to endpoint '{endpoint}' (requested cancellation: {cancellationToken.IsCancellationRequested})", e);
+
 			_discoveryCoordinator?.Post(new NodeMisbehaveMessage(endpoint, MisbehaviorType.FailedToConnect));
 			node?.DisconnectAsync();
 			return null;
